% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPart}
\alias{plotPart}
\title{Plot particle}
\usage{
plotPart(part, D = NULL, R = NULL, xlim = NULL, ylim = NULL,
  add = FALSE, edges = TRUE, col.edge = "red", vertices = TRUE,
  lines = FALSE, col.line = "black", lty.line = "dashed", ...)
}
\description{
Plot particle
}

